k/groups: moved recovery consumer to separate file
