+++ src/v/kafka/CMakeLists.txt
@@ -59,6 +59,7 @@ v_cc_library(
     server/fetch_session_cache.cc
     server/replicated_partition.cc
     server/partition_proxy.cc
+    server/group_recovery_consumer.cc
  DEPS
     Seastar::seastar
     v::bytes
